requirements.txt
Django>=4.2
django-allauth>=0.57.0
python-dotenv
requests
PyJWT
cryptography
djangorestframework
numpy
Pillow
scikit-learn

#pip install -r requirements.txt